travis: build with Go 1.11.4
Go 1.11 automatically checks the GOFLAGS environment variable, so no
need to manually add it to invocations of the Go tools.

Use 1.11.4 specifically because the calculation of go.sum checksums
changed in that version (https://github.com/golang/go/issues/29191),
so using this version means that the checksums are compatible for
Go >= 1.11.4.